Rename event stores to start using the new one
