+++ lib/sandthorn_driver_sequel.rb
@@ -5,7 +5,7 @@ require "sandthorn_driver_sequel/event"
 require "sandthorn_driver_sequel/access"
 require "sandthorn_driver_sequel/storage"
 require 'sandthorn_driver_sequel/event_store'
-require 'sandthorn_driver_sequel/new_event_store'
+require 'sandthorn_driver_sequel/event_store'
 require 'sandthorn_driver_sequel/errors'
 require 'sandthorn_driver_sequel/migration'
